+++ Runtime/Components/Editor/SpriteAnimatorEditor.cs
@@ -3,6 +3,7 @@ using UnityEditor;
 
 namespace NekoLib.Components
 {
+    [CanEditMultipleObjects]
     [CustomEditor(typeof(SpriteAnimator))]
     public class SpriteAnimatorEditor : SpriteAnimatorEditorBase
     {